chore: add DataManager, SessionManager type
