+++ SignUp/SignUp/ValidateManager.swift
@@ -9,48 +9,33 @@ import Foundation
 
 class ValidateManager {
     enum ValidateError: Error {
-        case alreadyUsing
+        case alreadyUsingId
     }
     
     static var url = URL(string: "https://8r6ruzgzve.execute-api.ap-northeast-2.amazonaws.com/default/SwiftCamp")!
     
     static func validate(for id: String, completion: @escaping (Result<String, Error>) -> Void) {
         var req = URLRequest(url: url)
-        let bodyData = encode(json: ["id": id, "password": ""])
+        let bodyData = DataManager.encode(json: ["id": id, "password": ""])
         req.httpMethod = "POST"
         req.httpBody = bodyData
-        request(for: req, completion: { result in
+        SessionManager.request(for: req, completion: { result in
             switch result {
             case .success(let data):
-                guard let json = decode(data: data),
+                guard let json = DataManager.decode(data: data),
                       let status = json["status"] as? String else {
                     return
                 }
                 if status == "200" {
                     completion(.success("success"))
                 } else {
-                    completion(.failure(ValidateError.alreadyUsing))
+                    completion(.failure(ValidateError.alreadyUsingId))
                 }
             case .failure(_):
-                completion(.failure(ValidateError.alreadyUsing))
+                completion(.failure(ValidateError.alreadyUsingId))
             }
         })
     }
     
-    static func request(for request: URLRequest, completion: @escaping (Result<Data, Error>) -> Void) {
-        URLSession.shared.dataTask(with: request, completionHandler: { data, _, error in
-            guard let data = data, error == nil else {
-                return
-            }
-            completion(.success(data))
-        }).resume()
-    }
     
-    private static func encode(json: [String: String]) -> Data? {
-        return try? JSONSerialization.data(withJSONObject: json, options: [])
-    }
-    
-    private static func decode(data: Data) -> [String: Any]? {
-        return try? JSONSerialization.jsonObject(with: data, options: []) as? [String: Any]
-    }
 }
